+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add `aerolab upgrade [--edge]` command to allow upgrading to latest stable (or latest pre-release).
 * Windows: catch the use of command prompt and warn against it.
 * Improvement: replace the `mesh/multicast` fix system with `aeroconf` parser.
+* Improvement: in `aerolab conf namespace-memory`, adjust `data-size` instead if using aerospike `v7+`, and only if required. Otherwise `noop`.
 * TODO: Monitor agi instance states and react accordingly - sizing or cycling from spot to on-demand types.
 * TODO: Document agi instance state monitor.